\name{ex07.03}
\alias{ex07.03}
\alias{ex07.03}
\title{Zar Data Set: ex07.03}
\description{Zar Data Set: ex07.03}
\details{Zar Data Set: ex07.03}
\docType{data}
\format{The format is:
num [1:12] 0.2 -0.5 -1.3 -1.6 -0.7 0.4 -0.1 0 -0.6 -1.1 ...}
\references{Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th Edition.
Pearson Prentice-Hall. Upper Saddle River, NJ. ISBN-10: 0131008463.
ISBN-13: 978013100846.
\url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}}
\keyword{datasets}
\examples{demo(ex07.03)}

