\docType{data}
\name{ex08.08}
\alias{ex08.08}
\title{Zar Data Set: ex08.08}
\format{The format is: List of 2
\describe{
\item{\code{$Weight}}{num [1:10] 72.5 71.7 60.8 63.2 71.4 73.1
      77.9 75.7 72 69}
\item{\code{$Height}}{num [1:11] 183 172 180 190 191 ...}
}}
\description{
  Zar Data Set: ex08.08
}
\details{
  Zar Data Set: ex08.08
}
\examples{
demo(ex08.08)
}
\references{
  Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th
  Edition. Pearson Prentice-Hall. Upper Saddle River, NJ.
  ISBN-10: 0131008463. ISBN-13: 978013100846.
  \url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\keyword{datasets}

