\name{ex08.11}
\alias{ex08.11}
\alias{ex08.11}
\title{Zar Data Set: ex08.11}
\description{Zar Data Set: ex08.11}
\details{Zar Data Set: ex08.11}
\docType{data}
\format{The format is: List of 2
\describe{
\item{\code{$Males}}{num [1:7] 193 188 185 183 180 175 170}
\item{\code{$Females}}{num [1:5] 178 173 168 165 163}
}}
\references{Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th Edition.
Pearson Prentice-Hall. Upper Saddle River, NJ. ISBN-10: 0131008463.
ISBN-13: 978013100846.
\url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}}
\keyword{datasets}
\examples{demo(ex08.11)}

