\docType{data}
\name{ex10.03}
\alias{ex10.03}
\title{Zar Data Set: ex10.03}
\format{A data frame with 18 observations on the following 2 variables.
\describe{
\item{\code{Variety}}{a factor with 3 levels: \code{G}, \code{A}, and \code{L}}
\item{\code{K}}{a numeric vector}
}}
\description{
  Zar Data Set: ex10.03
}
\details{
  Zar Data Set: ex10.03
}
\examples{
demo(ex10.03)
}
\references{
  Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th
  Edition. Pearson Prentice-Hall. Upper Saddle River, NJ.
  ISBN-10: 0131008463. ISBN-13: 978013100846.
  \url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\keyword{datasets}

