% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zar5-package.R
\docType{data}
\name{ex12.04}
\alias{ex12.04}
\title{Zar Data Set: ex12.04}
\format{A data frame with 15 observations on the following 3 
variables.
\describe{
\item{\code{Block}}{a numeric vector coding for block}
\item{\code{Treatment}}{a numeric vector coding for treatment}
\item{\code{Time}}{a numeric vector for the time to anesthesia (min.)}
}}
\description{
Zar Data Set: ex12.04
}
\details{
Zar Data Set: ex12.04
}
\examples{

demo(ex12.04)

}
\references{
Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th
  Edition. Pearson Prentice-Hall. Upper Saddle River, NJ. ISBN-10:
  0131008463. ISBN-13: 978013100846.
\url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\keyword{datasets}
