\docType{data}
\name{ex16.01}
\alias{ex16.01}
\title{Zar Data Set: ex16.01}
\format{A data frame with 20 observations on the following 3 variables.
\describe{
\item{\code{fat}}{a numeric vector}
\item{\code{lean}}{a numeric vector}
\item{\code{month}}{a factor with levels \code{Dec},
\code{Feb}, \code{Jan}, and \code{Mar}}
}}
\description{
  Zar Data Set: ex16.01
}
\details{
  Zar Data Set: ex16.01
}
\examples{
demo(ex16.01)
}
\references{
  Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th
  Edition. Pearson Prentice-Hall. Upper Saddle River, NJ.
  ISBN-10: 0131008463. ISBN-13: 978013100846.
  \url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\keyword{datasets}

