% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.tailed.variance.ratio.test.R
\name{two_tailed_variance_ratio_test}
\alias{two_tailed_variance_ratio_test}
\title{Two tailed variance ratio test}
\usage{
two_tailed_variance_ratio_test(group_1, group_2, alpha)
}
\arguments{
\item{group_1}{a numeric vector}

\item{group_2}{a numeric vector}

\item{alpha}{the desired significance cut off}
}
\description{
Two tailed variance ratio test
}
\details{
The two tailed variance ratio test is intended for testing for difference between the variances of two groups.
Refer to Biostatistical Analysis, by Jerrold H. Zar, fifth edition, section 8.5, page 151
}
\references{
Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th Edition. Pearson
  Prentice-Hall. Upper Saddle River, NJ. ISBN-10: 0131008463. ISBN-13:
  978013100846.
  \url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\author{
Daniel Kick (\email{drk8b9@mail.missouri.edu})
}
