% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zar5-package.R
\docType{package}
\name{Zar5-package}
\alias{Zar5-package}
\alias{Zar5}
\title{Zar5}
\description{
Data sets and example code for Zar's Biostatistical Analysis (5th 
Edition).
}
\details{
\tabular{ll}{
Package: \tab Zar5\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2018-02-27\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\references{
Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th 
  Edition. Pearson Prentice-Hall. Upper Saddle River, NJ. ISBN-10:
  0131008463. ISBN-13: 978013100846.
\url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\author{
Kevin Middleton (\email{middletonk@missouri.edu})
}
\keyword{package}
