\name{ex10.01}
\alias{ex10.01}
\alias{ex10.01}
\title{Zar Data Set: ex10...}
\description{Zar Data Set: ex10.01}
\details{Zar Data Set: ex10.01}
\docType{data}
\references{Zar, J.H. 2010. \emph{Biostatistical Analysis (5th Edition)}.
Pearson Prentice-Hall. Upper Saddle River, NJ. ISBN-10: 0131008463.
ISBN-13: 978013100846.
\url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}}
\keyword{datasets}
\examples{data(ex10.01)
str(ex10.01)

summary(aov(X ~ Feed, data = ex10.01))}

