\name{ex10.01}
\alias{ex10.01}
\alias{ex10.01}
\title{Zar Data Set: ex10.01}
\description{Zar Data Set: ex10.01}
\details{Zar Data Set: ex10.01}
\docType{data}
\format{A data frame with 20 observations on the following 2 variables.
\describe{
\item{\code{X}}{a numeric vector}
\item{\code{Feed}}{a factor with levels \code{1}, \code{2}, \code{3}, and \code{4}}
}}
\references{Zar, J.H. 2010. \emph{Biostatistical Analysis (5th Edition)}.
Pearson Prentice-Hall. Upper Saddle River, NJ. ISBN-10: 0131008463.
ISBN-13: 978013100846.
\url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}}
\keyword{datasets}
\examples{demo(ex10.01)}

