\docType{data}
\name{ex12.01}
\alias{ex12.01}
\title{Zar Data Set: ex12.01}
\format{A data frame with 20 observations on the following 3
variables.
\describe{
\item{\code{Ca}}{a numeric vector}
\item{\code{Trt}}{a factor with levels \code{No} and \code{Yes}}
\item{\code{Sex}}{a factor with levels \code{F} and \code{M}}
}}
\description{
  Zar Data Set: ex12.01
}
\details{
  Zar Data Set: ex12.01
}
\examples{
demo(ex12.01)
}
\references{
  Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th
  Edition. Pearson Prentice-Hall. Upper Saddle River, NJ.
  ISBN-10: 0131008463. ISBN-13: 978013100846.
  \url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\keyword{datasets}

