\docType{data}
\name{ex13.03}
\alias{ex13.03}
\title{Zar Data Set: ex13.03}
\format{A data frame with 5 observations on the following 4
variables.
\describe{
\item{\code{Group.1}}{a numeric vector}
\item{\code{Group.2}}{a numeric vector}
\item{\code{Group.3}}{a numeric vector}
\item{\code{Group.4}}{a numeric vector}
}}
\description{
  Zar Data Set: ex13.03
}
\details{
  Zar Data Set: ex13.03
}
\examples{
demo(ex13.03)
}
\references{
  Zar, J.H. 2010. \emph{Biostatistical Analysis}. 5th
  Edition. Pearson Prentice-Hall. Upper Saddle River, NJ.
  ISBN-10: 0131008463. ISBN-13: 978013100846.
  \url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}
}
\keyword{datasets}

