\name{Zar5-package}
\alias{Zar5-package}
\alias{Zar5-package}
\title{Zar5...}
\description{Zar5
Data sets and example code for Zar's Biostatistical Analysis (5th Edition).}
\details{\tabular{ll}{ 
Package: \tab Zar5\cr
Type: \tab Package\cr
Version: \tab 0.1-3\cr
Date: \tab 2011-01-30\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr }}
\alias{Zar5-package}
\alias{Zar5}
\docType{package}
\author{Kevin Middleton (\email{kmm@csusb.edu})}
\references{Zar, J.H. 2010. \emph{Biostatistical Analysis (5th Edition)}.
Pearson Prentice-Hall. Upper Saddle River, NJ. ISBN-10: 0131008463.
ISBN-13: 978013100846.
\url{http://www.pearsonhighered.com/educator/product/Biostatistical-Analysis/9780131008465.page}}
\keyword{package}

